progress

+++ hw7/main.go
@@ -7,11 +7,8 @@ import (
 	flag "github.com/spf13/pflag"
 )
 
-var offset *int64 = flag.Int64("offset", 0, "offset before copy")
-var limit *int64 = flag.Int64("limit", 0, "limit of copy size")
-
-// var fileFrom *string = flag.Int64("from", 0, "offset before copy")
-// var fileTo *string = flag.Int64("limit", 0, "limit of copy size")
+var offset *int = flag.Int("offset", 0, "offset before copy")
+var limit *int = flag.Int("limit", 0, "limit of copy size")
 
 func main() {
 	flag.Parse()
